Initial import.
I was going to use this for obfsNG, and the code initially lived in a
obfs4proxy branch, but "Post-quantum key exchange – a new hope" was
published during the design phase leading me to switch primitives.

I am releasing this separately in the hopes that it will be useful to
somebody.